move get_env_var_pair to environment
This is only used in environment, so it should live there too.
